+++ admin/src/web/controllers/users.py
@@ -1,5 +1,4 @@
-import string
-from flask import render_template,request, url_for, redirect
+from flask import abort, render_template,request, session, url_for, redirect
 from sqlalchemy import asc, desc
 from core import user
 from flask import Blueprint
@@ -8,15 +7,25 @@ from flask import flash
 from core.user.roles import Role
 from core.user.users import User
 from core.user import create_user, list_roles, list_users, search_users, update_user,delete_user
+from web.helpers.auth import check_permission, is_authenticated
 #from src.web.handlers.auth import login_required
 
 
 bprint = Blueprint("users", __name__, url_prefix="/usuarios")
 
 @bprint.get("/")
-#@login_required
-#@check("user_index")
+
 def index():
+    '''
+        Función que muestra las tablas de usuario, permite busqueda y paginación
+        Parameters: Ninguno(Depende en los parametros de la query)
+        Returns: Renderiza template HTML para la lista de usuarios con paginación   
+     '''
+    if not is_authenticated(session):
+        return abort(401)
+    
+    if not check_permission(session, "users_index"):
+        return abort(403)
     query = request.args.get('query',"")
     role = request.args.get('role', None)
     active = request.args.get('active', None)
@@ -24,15 +33,12 @@ def index():
     sort_by = request.args.get('sort_by', 'email') 
     order = request.args.get('order', 'asc')
 
-    #Ya funciona
     if active == "True":
         active = True
     elif active == "False":
         active = False
     else:
         active = None
-
-    
 
     users = search_users(email=query, role=role, active=active, page=page,sort_by=sort_by, order=order,)
 
@@ -41,9 +47,12 @@ def index():
 
 @bprint.post("/activar_usuario")
 def activar_usuario():
-    """
-    Funcion que permite habilitar/deshabilitar el usuario a menos que sea System Admin
-    """
+    '''Funcion que permite habilitar/deshabilitar el usuario a menos que sea System Admin'''
+    if not is_authenticated(session):
+        return abort(401)
+    
+    if not check_permission(session, "users_activar_usuario"):
+        return abort(403)
     chosen_id = request.form['id']
     query = request.form['query']
     user = User.query.get(chosen_id)
@@ -62,9 +71,16 @@ def activar_usuario():
 
 @bprint.route("/edit_user/<int:user_id>", methods=["GET", "POST"])
 def edit_user(user_id):
-    """ Función que edita el usuario y agrega los parametros a un diccionario para ahorrar 
-        validación en la funcion update_user
-    """
+    ''' 
+    Función que edita el usuario y agrega los parametros a un diccionario para 
+    no tener que hacer validación extra en update_user
+    '''
+
+    if not is_authenticated(session):
+        return abort(401)
+    
+    if not check_permission(session, "users_edit_user"):
+        return abort(403)
     user = User.query.get_or_404(user_id)  
     roles = Role.query.all()  
 
@@ -86,6 +102,18 @@ def edit_user(user_id):
 
 @bprint.post("/delete_user")
 def delete_user_controller():
+    '''
+        Función que elimina fisicamente un usuario
+        Parameters: Ninguno(Depende en los parametros de la query)
+        Raise: ValueError propagado por delete_user() 
+     '''
+    if not is_authenticated(session):
+        return abort(401)
+    
+    if not check_permission(session, "users_delete_user_controller"):
+        return abort(403)
+    
+
     user_id = request.form.get("user_id")  
     try:
         delete_user(user_id) 
@@ -97,6 +125,19 @@ def delete_user_controller():
 
 @bprint.route("/register_user", methods=["GET", "POST"])
 def register_user():
+    '''
+        Función que registra un nuevo usuario
+        Parameters: Ninguno(Depende en los parametros de la query)
+        Raise: ValueError propagado de create_user()
+        Returns: Renderiza template HTML para el formulario de registro y al index si el registro es correcto  
+     '''
+
+    if not is_authenticated(session):
+        return abort(401)
+    
+    if not check_permission(session, "users_register_user"):
+        return abort(403)
+
     roles = list_roles()
     roles = list_roles()
